Refactoring event utils out of website-next, adding tests and functionality to handle full image URL and http to https conversion
