+++ package.json
@@ -24,6 +24,7 @@
     "copy-webpack-plugin": "^3.0.1",
     "css-loader": "^0.25.0",
     "css-modules-require-hook": "^4.0.3",
+    "date-fns": "^1.3.0",
     "deep-extend": "^0.4.1",
     "empty": "^0.10.1",
     "ent": "^2.2.0",
