+++ website-next/package.json
@@ -30,7 +30,6 @@
     "body-parser": "^1.15.1",
     "chalk": "^1.1.1",
     "classnames": "^2.2.3",
-    "date-fns": "^1.3.0",
     "dotenv": "^2.0.0",
     "express": "^4.13.4",
     "express-prismic": "^1.1.1",
